feat: just change type of id in long

+++ src/dao/CrudOperations.java
@@ -6,6 +6,6 @@ public interface CrudOperations<E> {
 
     List<E> saveAll(List<E> entities); // Create if entities not exists, update if entities exists
     List<E> getAll(int offset, int limit);
-    E findById(int E_id);
-    boolean deleteOperation(int id);
+    E findById(long E_id);
+    boolean deleteOperation(long id);
 }
